+++ homeassistant/components/monarch_money/strings.json
